UCT/SOCKCM: fix header length type

+++ src/uct/tcp/sockcm/sockcm_def.h
@@ -25,8 +25,8 @@ typedef struct uct_sockcm_iface   uct_sockcm_iface_t;
 typedef struct uct_sockcm_ep      uct_sockcm_ep_t;
 
 typedef struct uct_sockcm_priv_data_hdr {
-    unsigned int length;     /* length of the private data */
-    int8_t       status;
+    ssize_t length;
+    int8_t  status;
 } uct_sockcm_priv_data_hdr_t;
 
 typedef struct uct_sockcm_conn_param {
